refactor: unit test improvements
