+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [BREAKING] Updated keystore to accept arbitrarily large public keys (#833).
 * Added Examples to Mdbook for Web Client (#850).
 * Added account code to `miden account --show` command (#835).
+* Exposed new test utilities in the `testing` feature (#882).
 
 ## 0.8.2 (TBD)
