+++ Cargo.toml
@@ -19,10 +19,11 @@ repository = "https://github.com/0xPolygonMiden/miden-client"
 
 [workspace.dependencies]
 async-trait = "0.1"
-miden-lib = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "next", default-features = false }
-miden-objects = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "next", default-features = false }
-miden-tx = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "next", default-features = false, features = ["async"] }
-miden-proving-service-client = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "next",  default-features = false, features = ["tx-prover"] }
+miden-lib = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "tomyrd-mock-private-notes", default-features = false }
+miden-objects = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "tomyrd-mock-private-notes", default-features = false }
+miden-test = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "tomyrd-mock-private-notes", default-features = false }
+miden-tx = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "tomyrd-mock-private-notes", default-features = false, features = ["async"] }
+miden-proving-service-client = { git = "https://github.com/0xPolygonMiden/miden-base", branch = "tomyrd-mock-private-notes", default-features = false, features = ["tx-prover"] }
 miette = { version = "7.2", features = ["fancy"] }
 rand = { version = "0.9" }
 serde = { version = "1.0", features = ["derive"] }
