+++ crates/rust-client/Cargo.toml
@@ -24,7 +24,7 @@ default = ["std", "tonic/channel"]
 idxdb = ["dep:base64", "dep:serde-wasm-bindgen", "dep:wasm-bindgen", "dep:wasm-bindgen-futures", "dep:serde", "dep:getrandom"]
 sqlite = ["dep:rusqlite", "dep:deadpool-sqlite", "std"]
 std = ["miden-objects/std","miden-proving-service-client/std"]
-testing = ["miden-objects/testing", "miden-lib/testing", "miden-tx/testing"]
+testing = ["miden-objects/testing", "miden-lib/testing", "miden-tx/testing", "dep:miden-test", "dep:uuid", "dep:toml"]
 tonic = ["std", "tonic/transport", "tonic/tls-ring", "tonic/tls-native-roots"]
 web-tonic = ["dep:tonic-web-wasm-client", "dep:getrandom"]
 
@@ -37,6 +37,7 @@ hex = { version = "0.4" }
 miden-proving-service-client = { workspace = true , features = ["tx-prover"]}
 miden-lib = { workspace = true }
 miden-objects = { workspace = true }
+miden-test = { workspace = true, features = ["async"], optional = true}
 miden-tx = { workspace = true, features = ["async"] }
 prost = { version = "0.13", default-features = false, features = ["derive"] }
 rand = { workspace = true }
@@ -50,6 +51,8 @@ tracing = { workspace = true }
 wasm-bindgen = { version = "0.2", features = ["serde-serialize"], optional = true }
 wasm-bindgen-futures = { version = "0.4", optional = true }
 getrandom = { version = "0.3", optional = true, features = ["wasm_js"] }
+uuid = { version = "1.10", features = ["serde", "v4", "js"], optional = true }
+toml = { version = "0.8", optional = true }
 
 [package.metadata.cargo-machete]
 ignored = ["getrandom"]
@@ -62,7 +65,6 @@ miden-client = { path = ".", features = ["testing", "concurrent", "sqlite", "ton
 miden-lib = { workspace = true, features = ["testing"] }
 miden-objects = { workspace = true, default-features = false, features = ["testing"] }
 tokio = { workspace = true }
-uuid = { version = "1.10", features = ["serde", "v4"] }
 web-sys = { version = "0.3", features = ["console"]}
 
 [build-dependencies]
