+++ crates/rust-client/src/lib.rs
@@ -113,8 +113,8 @@ pub mod transaction;
 #[cfg(feature = "std")]
 pub mod builder;
 
-#[cfg(test)]
-pub mod mock;
+#[cfg(feature = "testing")]
+mod test_utils;
 
 #[cfg(test)]
 pub mod tests;
@@ -184,6 +184,8 @@ pub mod utils {
 #[cfg(feature = "testing")]
 pub mod testing {
     pub use miden_objects::testing::*;
+
+    pub use crate::test_utils::*;
 }
 
 use alloc::sync::Arc;
