+++ crates/rust-client/src/test_utils/mod.rs
@@ -1,0 +1,4 @@
+pub mod mock;
+
+#[cfg(feature = "std")]
+pub mod common;
