+++ tests/Cargo.toml
@@ -20,8 +20,6 @@ miden-client = { path = "../crates/rust-client", features = [
 ] }
 miden-objects = { workspace = true, default-features = false, features = ["testing"] }
 rand = { workspace = true }
-toml = { version = "0.8" }
-uuid = { version = "1.10", features = ["serde", "v4"] }
 
 [dev-dependencies]
 async-trait = { version = "0.1" }
@@ -32,6 +30,3 @@ winter-maybe-async = { version = "0.12", features = ["async"]}
 [[test]]
 name = "integration"
 path = "integration/main.rs"
-
-[lib]
-path = "integration/lib.rs"
