--- tests/integration/lib.rs
@@ -1,2 +1,0 @@
-/// Provides utility functions for testing
-pub mod common;
