+++ tests/integration/main.rs
@@ -2,39 +2,27 @@ use std::{sync::Arc, time::Duration};
 
 use miden_client::{
     ClientError, ONE,
-    account::Account,
     builder::ClientBuilder,
-    note::NoteRelevance,
     rpc::{Endpoint, NodeRpcClient, TonicRpcClient, domain::account::AccountDetails},
-    store::{
-        InputNoteRecord, InputNoteState, NoteFilter, OutputNoteState, TransactionFilter,
-        input_note_states::ConsumedAuthenticatedLocalNoteState,
-    },
-    sync::{NoteTagSource, TX_GRACEFUL_BLOCKS},
+    store::{InputNoteRecord, InputNoteState, NoteFilter, OutputNoteState, TransactionFilter},
+    testing::common::*,
     transaction::{
-        PaymentTransactionData, TransactionExecutorError, TransactionProver,
-        TransactionProverError, TransactionRequestBuilder, TransactionStatus,
+        PaymentTransactionData, TransactionProver, TransactionProverError,
+        TransactionRequestBuilder, TransactionStatus,
     },
 };
 use miden_objects::{
-    account::{AccountId, AccountStorageMode},
+    account::AccountStorageMode,
     asset::{Asset, FungibleAsset},
     note::{NoteFile, NoteType},
     transaction::{ProvenTransaction, ToInputNoteCommitments, TransactionWitness},
 };
-
-mod common;
-use common::*;
 use winter_maybe_async::maybe_async_trait;
 
 mod custom_transactions_tests;
 mod fpi_tests;
 mod onchain_tests;
 mod swap_transactions_tests;
-
-/// Constant that represents the number of blocks until the p2idr can be recalled. If this value is
-/// too low, some tests might fail due to expected recall failures not happening.
-const RECALL_HEIGHT_DELTA: u32 = 50;
 
 #[tokio::test]
 async fn test_client_builder_initializes_client_with_endpoint() -> Result<(), ClientError> {
@@ -94,38 +82,6 @@ async fn test_client_builder_fails_without_keystore() {
         .await;
 
     assert!(result.is_err(), "Expected client build to fail without a keystore");
-}
-
-#[tokio::test]
-async fn test_added_notes() {
-    let (mut client, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
-
-    let faucet_account_header =
-        insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap()
-            .0;
-
-    // Mint some asset for an account not tracked by the client. It should not be stored as an
-    // input note afterwards since it is not being tracked by the client
-    let fungible_asset = FungibleAsset::new(faucet_account_header.id(), MINT_AMOUNT).unwrap();
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
-        fungible_asset,
-        AccountId::try_from(ACCOUNT_ID_REGULAR).unwrap(),
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-    println!("Running Mint tx...");
-    execute_tx_and_sync(&mut client, faucet_account_header.id(), tx_request).await;
-
-    // Check that no new notes were added
-    println!("Fetching Committed Notes...");
-    let notes = client.get_input_notes(NoteFilter::Committed).await.unwrap();
-    assert!(notes.is_empty())
 }
 
 #[tokio::test]
@@ -218,465 +174,6 @@ async fn test_multiple_tx_on_same_block() {
         sender_account.account().vault().get_balance(faucet_account_id).unwrap(),
         MINT_AMOUNT - (TRANSFER_AMOUNT * 2)
     );
-}
-
-#[tokio::test]
-async fn test_p2id_transfer() {
-    let (mut client, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
-
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    // First Mint necessary token
-    mint_and_consume(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
-    assert_account_has_single_asset(&client, from_account_id, faucet_account_id, MINT_AMOUNT).await;
-
-    // Do a transfer from first account to second account
-    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2ID tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
-        None,
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-
-    let note = tx_request.expected_output_notes().next().unwrap().clone();
-    let transaction_id = execute_tx(&mut client, from_account_id, tx_request).await;
-
-    // Check that a note tag started being tracked for this note.
-    assert!(
-        client
-            .get_note_tags()
-            .await
-            .unwrap()
-            .into_iter()
-            .any(|tag| tag.source == NoteTagSource::Note(note.id()))
-    );
-
-    wait_for_tx(&mut client, transaction_id).await;
-
-    // Check that the tag is not longer being tracked
-    assert!(
-        !client
-            .get_note_tags()
-            .await
-            .unwrap()
-            .into_iter()
-            .any(|tag| tag.source == NoteTagSource::Note(note.id()))
-    );
-
-    // Check that note is committed for the second account to consume
-    println!("Fetching Committed Notes...");
-    let notes = client.get_input_notes(NoteFilter::Committed).await.unwrap();
-    assert!(!notes.is_empty());
-
-    // Consume P2ID note
-    println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
-    execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
-
-    // Ensure we have nothing else to consume
-    let current_notes = client.get_input_notes(NoteFilter::Committed).await.unwrap();
-    assert!(current_notes.is_empty());
-
-    let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
-    let seed = regular_account.seed().cloned();
-    let regular_account: Account = regular_account.into();
-
-    // The seed should not be retrieved due to the account not being new
-    assert!(!regular_account.is_new() && seed.is_none());
-    assert_eq!(regular_account.vault().assets().count(), 1);
-    let asset = regular_account.vault().assets().next().unwrap();
-
-    // Validate the transfered amounts
-    if let Asset::Fungible(fungible_asset) = asset {
-        assert_eq!(fungible_asset.amount(), MINT_AMOUNT - TRANSFER_AMOUNT);
-    } else {
-        panic!("Error: Account should have a fungible asset");
-    }
-
-    let regular_account: Account = client.get_account(to_account_id).await.unwrap().unwrap().into();
-    assert_eq!(regular_account.vault().assets().count(), 1);
-    let asset = regular_account.vault().assets().next().unwrap();
-
-    if let Asset::Fungible(fungible_asset) = asset {
-        assert_eq!(fungible_asset.amount(), TRANSFER_AMOUNT);
-    } else {
-        panic!("Error: Account should have a fungible asset");
-    }
-
-    assert_note_cannot_be_consumed_twice(&mut client, to_account_id, notes[0].id()).await;
-}
-
-#[tokio::test]
-async fn test_p2id_transfer_failing_not_enough_balance() {
-    let (mut client, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
-
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    // First Mint necessary token
-    mint_and_consume(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
-
-    // Do a transfer from first account to second account
-    let asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT + 1).unwrap();
-    println!("Running P2ID tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
-        None,
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-    execute_failing_tx(
-        &mut client,
-        from_account_id,
-        tx_request,
-        ClientError::AssetError(miden_objects::AssetError::FungibleAssetAmountNotSufficient {
-            minuend: MINT_AMOUNT,
-            subtrahend: MINT_AMOUNT + 1,
-        }),
-    )
-    .await;
-}
-
-#[tokio::test]
-async fn test_p2idr_transfer_consumed_by_target() {
-    let (mut client, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
-
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    // First Mint necessary token
-    let note = mint_note(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
-    println!("about to consume");
-
-    //Check that the note is not consumed by the target account
-    assert!(matches!(
-        client.get_input_note(note.id()).await.unwrap().unwrap().state(),
-        InputNoteState::Committed { .. }
-    ));
-
-    consume_notes(&mut client, from_account_id, &[note.clone()]).await;
-    assert_account_has_single_asset(&client, from_account_id, faucet_account_id, MINT_AMOUNT).await;
-
-    // Check that the note is consumed by the target account
-    let input_note = client.get_input_note(note.id()).await.unwrap().unwrap();
-    assert!(matches!(input_note.state(), InputNoteState::ConsumedAuthenticatedLocal { .. }));
-    if let InputNoteState::ConsumedAuthenticatedLocal(ConsumedAuthenticatedLocalNoteState {
-        submission_data,
-        ..
-    }) = input_note.state()
-    {
-        assert_eq!(submission_data.consumer_account, from_account_id);
-    } else {
-        panic!("Note should be consumed");
-    }
-
-    // Do a transfer from first account to second account with Recall. In this situation we'll do
-    // the happy path where the `to_account_id` consumes the note
-    println!("getting balance");
-    let from_account_balance = client
-        .get_account(from_account_id)
-        .await
-        .unwrap()
-        .unwrap()
-        .account()
-        .vault()
-        .get_balance(faucet_account_id)
-        .unwrap_or(0);
-    let to_account_balance = client
-        .get_account(to_account_id)
-        .await
-        .unwrap()
-        .unwrap()
-        .account()
-        .vault()
-        .get_balance(faucet_account_id)
-        .unwrap_or(0);
-    let current_block_num = client.get_sync_height().await.unwrap();
-    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
-        Some(current_block_num + RECALL_HEIGHT_DELTA),
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-    execute_tx_and_sync(&mut client, from_account_id, tx_request.clone()).await;
-
-    // Check that note is committed for the second account to consume
-    println!("Fetching Committed Notes...");
-    let notes = client.get_input_notes(NoteFilter::Committed).await.unwrap();
-    assert!(!notes.is_empty());
-
-    // Make the `to_account_id` consume P2IDR note
-    let note_id = tx_request.expected_output_notes().next().unwrap().id();
-    println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build().unwrap();
-    execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
-    let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
-
-    // The seed should not be retrieved due to the account not being new
-    assert!(!regular_account.account().is_new() && regular_account.seed().is_none());
-    assert_eq!(regular_account.account().vault().assets().count(), 1);
-    let asset = regular_account.account().vault().assets().next().unwrap();
-
-    // Validate the transfered amounts
-    if let Asset::Fungible(fungible_asset) = asset {
-        assert_eq!(fungible_asset.amount(), from_account_balance - TRANSFER_AMOUNT);
-    } else {
-        panic!("Error: Account should have a fungible asset");
-    }
-
-    let regular_account: Account = client.get_account(to_account_id).await.unwrap().unwrap().into();
-    assert_eq!(regular_account.vault().assets().count(), 1);
-    let asset = regular_account.vault().assets().next().unwrap();
-
-    if let Asset::Fungible(fungible_asset) = asset {
-        assert_eq!(fungible_asset.amount(), to_account_balance + TRANSFER_AMOUNT);
-    } else {
-        panic!("Error: Account should have a fungible asset");
-    }
-
-    assert_note_cannot_be_consumed_twice(&mut client, to_account_id, note_id).await;
-}
-
-#[tokio::test]
-async fn test_p2idr_transfer_consumed_by_sender() {
-    let (mut client, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
-
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    // First Mint necessary token
-    mint_and_consume(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
-
-    // Do a transfer from first account to second account with Recall. In this situation we'll do
-    // the happy path where the `to_account_id` consumes the note
-    let from_account_balance = client
-        .get_account(from_account_id)
-        .await
-        .unwrap()
-        .unwrap()
-        .account()
-        .vault()
-        .get_balance(faucet_account_id)
-        .unwrap_or(0);
-    let current_block_num = client.get_sync_height().await.unwrap();
-    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
-        Some(current_block_num + RECALL_HEIGHT_DELTA),
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-    execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
-
-    // Check that note is committed
-    println!("Fetching Committed Notes...");
-    let notes = client.get_input_notes(NoteFilter::Committed).await.unwrap();
-    assert!(!notes.is_empty());
-
-    // Check that it's still too early to consume
-    println!("Consuming Note (too early)...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
-    let transaction_execution_result = client.new_transaction(from_account_id, tx_request).await;
-    assert!(transaction_execution_result.is_err_and(|err| {
-        matches!(
-            err,
-            ClientError::TransactionExecutorError(
-                TransactionExecutorError::TransactionProgramExecutionFailed(_)
-            )
-        )
-    }));
-
-    // Wait to consume with the sender account
-    println!("Waiting for note to be consumable by sender");
-    wait_for_blocks(&mut client, RECALL_HEIGHT_DELTA).await;
-
-    // Consume the note with the sender account
-    println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
-    execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
-
-    let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
-    // The seed should not be retrieved due to the account not being new
-    assert!(!regular_account.account().is_new() && regular_account.seed().is_none());
-    assert_eq!(regular_account.account().vault().assets().count(), 1);
-    let asset = regular_account.account().vault().assets().next().unwrap();
-
-    // Validate the sender hasn't lost funds
-    if let Asset::Fungible(fungible_asset) = asset {
-        assert_eq!(fungible_asset.amount(), from_account_balance);
-    } else {
-        panic!("Error: Account should have a fungible asset");
-    }
-
-    let regular_account: Account = client.get_account(to_account_id).await.unwrap().unwrap().into();
-    assert_eq!(regular_account.vault().assets().count(), 0);
-
-    // Check that the target can't consume the note anymore
-    assert_note_cannot_be_consumed_twice(&mut client, to_account_id, notes[0].id()).await;
-}
-
-#[tokio::test]
-async fn test_get_consumable_notes() {
-    let (mut client, authenticator) = create_test_client().await;
-
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    //No consumable notes initially
-    assert!(client.get_consumable_notes(None).await.unwrap().is_empty());
-
-    // First Mint necessary token
-    let note = mint_note(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
-
-    // Check that note is consumable by the account that minted
-    assert!(!client.get_consumable_notes(None).await.unwrap().is_empty());
-    assert!(!client.get_consumable_notes(Some(from_account_id)).await.unwrap().is_empty());
-    assert!(client.get_consumable_notes(Some(to_account_id)).await.unwrap().is_empty());
-
-    consume_notes(&mut client, from_account_id, &[note]).await;
-
-    //After consuming there are no more consumable notes
-    assert!(client.get_consumable_notes(None).await.unwrap().is_empty());
-
-    // Do a transfer from first account to second account
-    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
-        Some(100.into()),
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-    execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
-
-    // Check that note is consumable by both accounts
-    let consumable_notes = client.get_consumable_notes(None).await.unwrap();
-    let relevant_accounts = &consumable_notes.first().unwrap().1;
-    assert_eq!(relevant_accounts.len(), 2);
-    assert!(!client.get_consumable_notes(Some(from_account_id)).await.unwrap().is_empty());
-    assert!(!client.get_consumable_notes(Some(to_account_id)).await.unwrap().is_empty());
-
-    // Check that the note is only consumable after block 100 for the account that sent the
-    // transaction
-    let from_account_relevance = relevant_accounts
-        .iter()
-        .find(|relevance| relevance.0 == from_account_id)
-        .unwrap()
-        .1;
-    assert_eq!(from_account_relevance, NoteRelevance::After(100));
-
-    // Check that the note is always consumable for the account that received the transaction
-    let to_account_relevance = relevant_accounts
-        .iter()
-        .find(|relevance| relevance.0 == to_account_id)
-        .unwrap()
-        .1;
-    assert_eq!(to_account_relevance, NoteRelevance::Always);
-}
-
-#[tokio::test]
-async fn test_get_output_notes() {
-    let (mut client, authenticator) = create_test_client().await;
-
-    let (first_regular_account, faucet_account_header) =
-        setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
-
-    let from_account_id = first_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-    let random_account_id = AccountId::try_from(ACCOUNT_ID_REGULAR).unwrap();
-
-    // No output notes initially
-    assert!(client.get_output_notes(NoteFilter::All).await.unwrap().is_empty());
-
-    // First Mint necessary token
-    let note = mint_note(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
-
-    // Check that there was an output note but it wasn't consumed
-    assert!(client.get_output_notes(NoteFilter::Consumed).await.unwrap().is_empty());
-    assert!(!client.get_output_notes(NoteFilter::All).await.unwrap().is_empty());
-
-    consume_notes(&mut client, from_account_id, &[note]).await;
-
-    //After consuming, the note is returned when using the [NoteFilter::Consumed] filter
-    assert!(!client.get_output_notes(NoteFilter::Consumed).await.unwrap().is_empty());
-
-    // Do a transfer from first account to second account
-    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2ID tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(
-            vec![Asset::Fungible(asset)],
-            from_account_id,
-            random_account_id,
-        ),
-        None,
-        NoteType::Private,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-
-    let output_note_id = tx_request.expected_output_notes().next().unwrap().id();
-
-    // Before executing, the output note is not found
-    assert!(client.get_output_note(output_note_id).await.unwrap().is_none());
-
-    execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
-
-    // After executing, the note is only found in output notes
-    assert!(client.get_output_note(output_note_id).await.unwrap().is_some());
-    assert!(client.get_input_note(output_note_id).await.unwrap().is_none());
 }
 
 #[tokio::test]
@@ -1126,7 +623,7 @@ async fn test_consume_multiple_expected_notes() {
     let fungible_asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     let mint_tx_request = mint_multiple_fungible_asset(
         fungible_asset,
-        vec![to_account_ids[0], to_account_ids[0], to_account_ids[1], to_account_ids[1]],
+        &[to_account_ids[0], to_account_ids[0], to_account_ids[1], to_account_ids[1]],
         NoteType::Private,
         client.rng(),
     );
@@ -1632,89 +1129,4 @@ async fn test_unused_rpc_api() {
 
     assert_eq!(account_delta.nonce(), Some(ONE));
     assert_eq!(*account_delta.vault().fungible().iter().next().unwrap().1, MINT_AMOUNT as i64);
-}
-
-#[tokio::test]
-async fn test_stale_transactions_discarded() {
-    let (mut client, authenticator) = create_test_client().await;
-    let (regular_account, faucet_account_header) =
-        setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
-
-    let account_id = regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    // Mint a note
-    let note = mint_note(&mut client, account_id, faucet_account_id, NoteType::Private).await;
-    consume_notes(&mut client, account_id, &[note]).await;
-
-    // Create a transaction but don't submit it to the node
-    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-
-    let tx_request = TransactionRequestBuilder::pay_to_id(
-        PaymentTransactionData::new(vec![Asset::Fungible(asset)], account_id, account_id),
-        None,
-        NoteType::Public,
-        client.rng(),
-    )
-    .unwrap()
-    .build()
-    .unwrap();
-
-    // Execute the transaction but don't submit it to the node
-    let tx_result = client.new_transaction(account_id, tx_request).await.unwrap();
-    let tx_id = tx_result.executed_transaction().id();
-    client.testing_prove_transaction(&tx_result).await.unwrap();
-
-    // Store the account state before applying the transaction
-    let account_before_tx = client.get_account(account_id).await.unwrap().unwrap();
-    let account_commitment_before_tx = account_before_tx.account().commitment();
-
-    // Apply the transaction
-    client.testing_apply_transaction(tx_result).await.unwrap();
-
-    // Check that the account state has changed after applying the transaction
-    let account_after_tx = client.get_account(account_id).await.unwrap().unwrap();
-    let account_commitment_after_tx = account_after_tx.account().commitment();
-
-    assert_ne!(
-        account_commitment_before_tx, account_commitment_after_tx,
-        "Account commitment should change after applying the transaction"
-    );
-
-    // Verify the transaction is in pending state
-    let tx_record = client
-        .get_transactions(TransactionFilter::All)
-        .await
-        .unwrap()
-        .into_iter()
-        .find(|tx| tx.id == tx_id)
-        .unwrap();
-    assert!(matches!(tx_record.transaction_status, TransactionStatus::Pending));
-
-    // Sync the state, which should discard the old pending transaction
-    wait_for_blocks(&mut client, TX_GRACEFUL_BLOCKS + 1).await;
-
-    // Verify the transaction is now discarded
-    let tx_record = client
-        .get_transactions(TransactionFilter::All)
-        .await
-        .unwrap()
-        .into_iter()
-        .find(|tx| tx.id == tx_id)
-        .unwrap();
-
-    assert!(matches!(tx_record.transaction_status, TransactionStatus::Discarded));
-
-    // Check that the account state has been rolled back after the transaction was discarded
-    let account_after_sync = client.get_account(account_id).await.unwrap().unwrap();
-    let account_commitment_after_sync = account_after_sync.account().commitment();
-
-    assert_ne!(
-        account_commitment_after_sync, account_commitment_after_tx,
-        "Account commitment should change after transaction was discarded"
-    );
-    assert_eq!(
-        account_commitment_after_sync, account_commitment_before_tx,
-        "Account commitment should be rolled back to the value before the transaction"
-    );
 }
